fix(e2e): seed acceptances per user, and stop pinning which layer refuses a write
Two failures, one of them a real regression.

existing-user-second-company.spec.ts provisions its OWN user, so global-setup's
seeding never reached them -- the gate then disabled the Join button and the
spec's click timed out with no visible cause. That is the failure mode the plan
predicted for the suite generally, and a per-spec fix would leave the next spec
to rediscover it. So the seeding moves to e2e/fixtures/terms.ts with a header
saying to call it for every user a spec creates, and global-setup uses the same
helper rather than a private copy.

The other repeats a mistake CI already taught me on the Python side: the e2e
spec asserted the append-only TRIGGER's message, but over PostgREST the GRANT
refuses first with 42501, because service_role holds SELECT and INSERT only. It
now asserts the write was refused and the row is unchanged, without pinning
which of the two layers did it -- requiring the inner one makes the test fail
precisely when the outer one works.
